% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buzzMYcont.R
\name{buzzMYcont}
\alias{buzzMYcont}
\title{Fit an exploratory Bayesian mediation model.}
\usage{
buzzMYcont(
  dataset,
  X,
  M,
  Y,
  shape_m = NULL,
  rate_m = NULL,
  shape_y = NULL,
  rate_y = NULL,
  shape_a = NULL,
  rate_a = NULL,
  shape_b = NULL,
  rate_b = NULL,
  alpha_ind = NULL,
  beta_ind = NULL,
  tau_cprime = NULL,
  prec.m = NULL,
  prec.y = NULL,
  c.prime = NULL,
  taua = NULL,
  taub = NULL,
  ind.p = NULL,
  n_burnin = NULL,
  n_iter = NULL,
  thin = NULL
)
}
\arguments{
\item{dataset}{A data.frame containing the outcome, predictors, and mediators.}

\item{X}{A string or character vector giving the name(s) of the predictor
variable(s) in \code{dataset}.}

\item{M}{A character vector giving the names of mediator variables
in \code{dataset}.}

\item{Y}{A string giving the name of the outcome variable in \code{dataset}.}

\item{shape_m}{Numeric. Shape parameter for the gamma prior on mediator
residual precisions (\code{prec.m[j]}). If NULL, a dgamma(1, 0.001)
prior is used (default).}

\item{rate_m}{Numeric. Rate parameter for the gamma prior on mediator
residual precisions. If NULL, a dgamma(1, 0.001) prior is used (default).}

\item{shape_y}{Numeric. Shape parameter for the gamma prior on the outcome
residual precision (\code{prec.y}). If NULL, a dgamma(1, 0.001)
prior is used (default).}

\item{rate_y}{Numeric. Rate parameter for the gamma prior on the outcome
residual precision. If NULL, a dgamma(1, 0.001) prior is used (default).}

\item{shape_a}{Numeric. Shape parameter for the gamma prior on the slab
precision of the \eqn{a} paths. If NULL, a dgamma(1, 0.001)
prior is used (default).}

\item{rate_a}{Numeric. Rate parameter for the gamma prior on the slab
precision of the \eqn{a} paths. If NULL, a dgamma(1, 0.001)
prior is used (default).}

\item{shape_b}{Numeric. Shape parameter for the gamma prior on the slab
precision of the \eqn{b} paths. If NULL, a dgamma(1, 0.001)
prior is used (default).}

\item{rate_b}{Numeric. Rate parameter for the gamma prior on the slab
precision of the \eqn{b} paths. If NULL, a dgamma(1, 0.001)
prior is used (default).}

\item{alpha_ind}{Numeric. Alpha parameter for the beta prior on the inclusion
probability (\code{ind.p}). If NULL, a Beta(3, 3) prior is used (default).}

\item{beta_ind}{Numeric. Beta parameter for the beta prior on the inclusion
probability. If NULL, a Beta(3, 3) prior is used (default).}

\item{tau_cprime}{Numeric. Precision of the normal prior on the direct effects
of predictors (\code{c.prime}). If NULL, a dnorm(0, 1.0E-6)
prior is used (default).}

\item{prec.m}{Numeric. Initial value for the for the direct effect.
If NULL, initialized at 1 (default).}

\item{prec.y}{Numeric. Initial value for the outcome residual precision.
If NULL, initialized at 1 (default).}

\item{c.prime}{Numeric. Initial value for the direct effects (\code{X -> Y}).
If NULL, initialized at 0 (default).}

\item{taua}{Numeric. Initial value for the slab precision of the \eqn{a} paths.
If NULL, initialized at 1 (default).}

\item{taub}{Numeric. Initial value for the slab precision of the \eqn{b} paths.
If NULL, initialized at 1 (default).}

\item{ind.p}{Numeric in (0, 1). Initial value for the inclusion probability.
If NULL, initialized at 0.5 (default).}

\item{n_burnin}{Integer. Number of burn-in iterations for the MCMC sampler.
If NULL, 1000 iterations are used (default).}

\item{n_iter}{Integer. Number of MCMC iterations to sample after burn-in.
If NULL, 10000 iterations are used (default).}

\item{thin}{Integer. Thinning interval for MCMC samples.
If NULL, no thinning is applied (default = 1).}
}
\value{
An object of class \code{mcmc.list} containing posterior samples from JAGS.
}
\description{
Fits an explanatory Bayesian mediation model with continuous mediators and
continuous dependent variables. The function prepares the data, constructs
the JAGS model, runs MCMC sampling, and returns posterior samples.
}
\details{
The function estimates mediating effects under Exploratory Bayesian Mediation Analysis.

Internally, this function calls \code{prepare_ebmed_data()},
\code{build_ebmed_model()}, \code{define_init_values()},
and \code{run_ebmed_jags()}.
}
